First stab
Plan is to rely on COPYUID response to know UIDs of messages in
destination folder.  Gmail will no longer delete/expunge messages to
archive, but will move them to All Mail (just like moving to Trash)
and use COPYUID.  This unifies all approaches and allows for a simple
undo command.

Questions remain if the undo stack is maintained by the client or if
the engine supplies a simple, one-operation "stack", which doesn't
merely simplify the implementation on the client side but also allows
for the Account object to directly manage UIDs; otherwise, the COPYUID
UIDs must be recorded in the database in order to generate valid
ImapDB.EmailIdentifiers.